[FIX] product_supplierinfo_for_customer: Crash when the price is computed for a template instead of variant TT27887
