+++ product_supplierinfo_for_customer/models/__init__.py
@@ -1,6 +1,7 @@
 from . import product_pricelist
 from . import product_customerinfo
 from . import product_supplierinfo
+from . import product_customerinfo_mixin
 from . import product_product
 from . import product_template
 from . import res_partner
